dev-requirements.txt
# All versions pinned exactly is preferred for complete hermetic testing.
faiss-cpu==1.8.0
numpy==1.26.4
pillow==10.3.0
pytest==8.2.0
pytest-mock==3.14.0
torch==2.3.0
torchvision==0.18.0
transformers==4.39.3